requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
requests>=2.32.3
huggingface-hub>=0.20.3 